[#48] Add `ContextMenu` + Fix Tree Shaking
Problem:
We need a context menu component for deeper interactions, for example,
when long-pressing avatar, it should show a dropdown.

Also in nested components (which consists of more than one component),
like the new one, there is a problem with tree shaking. It should be
exported by component, not as a wildcard.

Solution:
The new component has been added. Tree shaking was fixed.